Add static mapping for ipopt/cyipopt (#1539)
The cyipopt package on conda is published under two names on PyPI: ipopt and cyipopt. It also exports
both 'cyipopt' (preferred) and 'ipopt' (now deprecated) packages. I am not sure whether all these combinations
need to be listed or if some are redundant. I'm also not sure if having duplicates like this works.

Co-authored-by: Matthew R. Becker <beckermr@users.noreply.github.com>