Enable `CondaForgeYAMLCleanup` migrator for v1 recipes (#3765)
* Enable `CondaForgeYAMLCleanup` migrator for v1 recipes

Enable `CondaForgeYAMLCleanup` migrator for v1 recipes.  Since it works
on `conda-forge.yml` only, it does not require any changes.  So the most
of the changes are related to extending the test.

* Fix handling v1recipes with no `license_file`

Fix an exception when the tested v1 recipe has no `license_file` key:

E       TypeError: 'NoneType' object is not subscriptable

* Fix formatting

* Add schema version check to `MiniMigrator.filter()`

* Call superclass in `CondaForgeYAMLCleanup.filter()`

* Actually enable schema version 1 in `CondaForgeYAMLCleanup`

---------

Co-authored-by: Matthew R. Becker <beckermr@users.noreply.github.com>