Support v1 recipes in `CrossPythonMigrator`
Update `CrossPythonMigrator` for v1 recipes.  Unfortunately, this yields
a bit suboptimal output, as it does not combine conditions:

```yaml
    - if: build_platform != target_platform
      then: python
    - if: build_platform != target_platform
      then: cross-python_${{ target_platform }}
    - if: build_platform != target_platform
      then: cython
```

That said, addressing this would increase the divergency between v0
and v1 code in the migrator, and `conda-recipe-manager convert` suffers
from the same issue.  I think a better way forward would be to write
a dedicated v1 recipe migrator that combines adjacent conditions.

Part of bug #3642