Do not combine non-adjacent conditions in v1 recipes
Modify the v1 recipe condition combiner not to merge non-adjacent
conditions.  While this is not really a problem for dependencies (though
it could break original dependency grouping), it did lead to reordering
test commands that could lead to breaking the recipe.

The algorithm aims to be as simple as possible.  It iterates over lists
tail-to-head and folds an "if" into the preceding item if it has
the same condition.

Fixes #3933