feat: add field allowed_tag_globs to version_updates schema (#4918)
* Add field tag_whitelist_regexp to version_updates schema

The field, if specified, gives regular expression for tags
which are considered. Tags not matching the regular expression
are ignored.

This feature is needed for cccl-python-feedstock, where the
regexp would be set to 'python-*'. See gh-4913 for context and
discussion.

* Use glob-like patterns as understood by fnmatch.fnmatch instead of regexp

Also renamed field to allowed_tag_patterns, and permit it to be a list.
The list of patterns is understood that tag matching any of them
participates in version ordering to determine the next tag to upgrade to.

* feat: rename and centralize logic

* Update conda_forge_tick/config_schema.py

* fix: update schema

* fix: update schema

---------

Co-authored-by: beckermr <becker.mr@gmail.com>
Co-authored-by: Matthew R. Becker <beckermr@users.noreply.github.com>