Don't trust last-modified for PR older then 7 days
See #5150, GitHub "Last-Modified" seem to means that the PR itself has
not received any modifications; but it _may_ be conflicting now,
The etag seem better but is also a weak etag.

Therefore, if Last-Mofified is older than 7 days, skip it,
refresh the PR and replace the Last_modified with the date,
this prevent PRs which really have identical outputs because not
modified to get force-refreshed only once every weeks.

This only affects PRs where the status is passing, and merging is clean
as we don't really care for others (I think).

304 do count on API calls, so maybe we should just care less, and make
unconditional calls ?

I also added the ability to locally test on a single feedstock (it
works), and modified the --offline to also not try to close PRs.
Otherwise it fais as my GH TOKEN is read-only; well, I was debugging
with a RW token yesterday and eded up commenting on a feedstock I did
not meant to...